버전 지정 없이 업데이트 가능하도록 update_version 구현

+++ launcher_abler/ConfigSelector.py
@@ -31,7 +31,6 @@ class ConfigSelector:
         """
         Read config.ini & Extract ABLER and Launcher version.
         """
-
         with open(self.config, "r") as f:
             for line in f.readlines():
                 line = line.strip("\n")
@@ -43,30 +42,27 @@ class ConfigSelector:
 
                 self.config_data.append(line)
 
+        if args.print:
+            print(f"[config.ini]")
+            print(f"ABLER    ver : {self.abler_ver}")
+            print(f"Launcher ver : {self.launcher_ver}")
+
         return
 
     def set_config(self):
         """
         Set ABLER & Launcher version.
         """
-        get_abler = self.abler_ver if args.abler is None else args.abler
-        get_launcher = self.launcher_ver if args.launcher is None else args.launcher
-
-        if args.print:
-            print(f"[config.ini]")
-            print(f"ABLER    ver : {get_abler}")
-            print(f"Launcher ver : {get_launcher}")
-
         with open(self.config, "w") as f:
             for line in self.config_data:
                 if "installed" in line:
                     line = line.split(" ")[:-1]
-                    line.append(get_abler)
+                    line.append(self.abler_ver)
                     line = " ".join(line)
 
                 elif "launcher" in line:
                     line = line.split(" ")[:-1]
-                    line.append(get_launcher)
+                    line.append(self.launcher_ver)
                     line = " ".join(line)
 
                 f.writelines(line)
@@ -78,7 +74,10 @@ class ConfigSelector:
         """
         Save back-up of config.ini
         """
-        shutil.copyfile(self.config, self.config_bak)
+        print("Copy config.ini to config.bak")
+
+        if not os.path.isfile(self.config_bak):
+            shutil.copyfile(self.config, self.config_bak)
 
         return
 
@@ -86,6 +85,8 @@ class ConfigSelector:
         """
         Reset back-up to config.ini
         """
+        print("Reset config.bak to config.ini")
+
         if os.path.isfile(self.config_bak):
             shutil.copyfile(self.config_bak, self.config)
             os.remove(self.config_bak)
@@ -96,29 +97,53 @@ class ConfigSelector:
         """
         Remove config.bak
         """
+        print("Remove config.bak")
+
         if os.path.isfile(self.config_bak):
             os.remove(self.config_bak)
 
         return
 
+    def update_version(self):
+        """
+        Auto update with args
+        """
+        if "abler" in args.update:
+            self.abler_ver = "0.0.1"
+
+        if "launcher" in args.update:
+            self.launcher_ver = "0.0.1"
+
 
 def main():
     config = ConfigSelector()
 
-    if args.copy:
-        config.copy_config()
+    if not os.path.isfile(config.config_bak):
+        if args.copy:
+            config.copy_config()
 
-    if args.reset:
-        config.reset_config()
+        else:
+            print("Please back-up config.ini with '--copy'")
 
-    if args.remove:
-        config.remove_backup()
+    else:
+        if args.copy:
+            print("Already exists back-up file.")
 
-    if args.run:
-        config.copy_config()
-        config.read_config()
-        config.set_config()
-        subprocess.call(config.launcher)
+        if args.reset:
+            config.reset_config()
+
+        if args.remove:
+            config.remove_backup()
+
+        if args.update:
+            config.copy_config()
+            config.read_config()
+            config.update_version()
+            config.set_config()
+
+        if args.run:
+            print(f"Run {config.launcher}")
+            subprocess.call(config.launcher)
 
     return
 
@@ -126,21 +151,23 @@ def main():
 if __name__ == "__main__":
     # Define parser
     parser = argparse.ArgumentParser(
-        description="Test ABLER launcher update", usage="Parser [options]"
+        description="ABLER launcher update test.", usage="Parser [options]"
     )
 
-    parser.add_argument("--launcher", "-l", action="store", default=None, type=str)
-    parser.add_argument("--abler", "-a", action="store", default=None, type=str)
+    parser.add_argument("--print", "-p", action="store_true")
     parser.add_argument("--copy", "-c", action="store_true")
     parser.add_argument("--reset", "-r", action="store_true")
     parser.add_argument("--remove", action="store_true")
-    parser.add_argument("--print", "-p", action="store_true")
+    parser.add_argument(
+        "--update", "-u", action="store", default=None, type=str, nargs="*"
+    )
     parser.add_argument("--run", action="store_true")
 
     args = parser.parse_args()
 
     if len(sys.argv) == 1:
-        print("Please set ABLER or Launcher version.")
+        print("ABLER launcher update test.")
+        print("Please set arguments.")
 
     else:
         main()
